Report circular requires at runtime
Circular requires often manifest themselves are cryptic "undefined is
not a function" errors. This occurs when you have circular requires and
partially exported modules are being returned by `require`.

To fix this, we can keep track of the stack of requires and mark them as
being completely finished exporting by popping them off the stack. By
keeping track of this, we can see where exactly the circular requires
are taking place.