Support for "use strict"
When running in strict mode, and executing a function with explicitly passing a context object, "this === null".
A number of modules (including underscore.js) expects that "this === window" at the global context. This holds true for node.js
Modulr looses this context when executing a factory. Pass the context explicitly.